refactor(managedSites): move doneHubService into providers subdirectory
Relocate doneHubService from root services to managedSites/providers
and update import path in managedSiteService.ts accordingly.
