+++ services/managedSites/managedSiteService.ts
@@ -22,8 +22,8 @@ import {
   getManagedSiteContext,
 } from "~/utils/managedSite"
 
-import * as doneHubService from "../doneHubService/doneHubService"
 import { userPreferences, type UserPreferences } from "../userPreferences"
+import * as doneHubService from "./providers/doneHubService"
 import * as newApiService from "./providers/newApi"
 import * as octopusService from "./providers/octopus"
 import * as veloeraService from "./providers/veloera"
